Remove multi-lifetime support from VisualStudioRefactorNotifyService
This class was an implicitly-active service that lived in configured-project scope when it had no initialization or shutdown logic. Make it a plain service.
