+++ src/Microsoft.VisualStudio.ProjectSystem.Managed.VS/ProjectSystem/VS/Refactor/VisualStudioRefactorNotifyService.cs
@@ -1,7 +1,9 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
+using System;
 using System.Collections.Generic;
 using System.ComponentModel.Composition;
+using System.Linq;
 using System.Threading.Tasks;
 
 using EnvDTE;
@@ -11,50 +13,124 @@ using Microsoft.VisualStudio.Shell.Interop;
 
 namespace Microsoft.VisualStudio.ProjectSystem.VS.Refactor
 {
-    [Export(typeof(IImplicitlyActiveService))]
     [Export(typeof(IRefactorNotifyService))]
     [AppliesTo(ProjectCapability.CSharpOrVisualBasic)]
-    internal sealed partial class VisualStudioRefactorNotifyService : AbstractMultiLifetimeComponent<VisualStudioRefactorNotifyService.Instance>, IImplicitlyActiveService, IRefactorNotifyService
+    internal class VisualStudioRefactorNotifyService : IRefactorNotifyService
     {
         private readonly IVsService<SDTE, DTE> _dte;
         private readonly IVsService<SVsSolution, IVsSolution> _solutionService;
 
         [ImportingConstructor]
-        public VisualStudioRefactorNotifyService(UnconfiguredProject unconfiguredProject,
-                                                 IProjectThreadingService threadingService,
-                                                 IVsService<SDTE, DTE> dte,
-                                                 IVsService<SVsSolution, IVsSolution> solutionService)
-            : base(threadingService.JoinableTaskContext)
+        public VisualStudioRefactorNotifyService(IVsService<SDTE, DTE> dte, IVsService<SVsSolution, IVsSolution> solutionService)
         {
             _dte = dte;
             _solutionService = solutionService;
         }
 
-        public Task ActivateAsync() => LoadAsync();
-        public Task DeactivateAsync() => UnloadAsync();
-
-        protected override Instance CreateInstance() => new Instance(_dte, _solutionService);
-
         public async Task OnBeforeGlobalSymbolRenamedAsync(string projectPath, IEnumerable<string> filePaths, string rqName, string newName)
         {
-            Requires.NotNull(projectPath, nameof(projectPath));
-            Requires.NotNull(filePaths, nameof(filePaths));
-            Requires.NotNull(rqName, nameof(rqName));
-            Requires.NotNull(newName, nameof(newName));
+            IVsHierarchy? projectHierarchy = await GetProjectHierarchy(projectPath);
+            if (projectHierarchy == null)
+            {
+                return;
+            }
 
-            Instance instance = await WaitForLoadedAsync();
-            await instance.OnBeforeGlobalSymbolRenamedAsync(projectPath, filePaths, rqName, newName);
+            if (!(projectHierarchy is IVsHierarchyRefactorNotify refactorNotify))
+            {
+                return;
+            }
+
+            uint[] ids = GetIdsForFiles(projectHierarchy, filePaths).ToArray();
+
+            refactorNotify.OnBeforeGlobalSymbolRenamed(cItemsAffected: (uint)ids.Length,
+                                                       rgItemsAffected: ids,
+                                                       cRQNames: (uint)1,
+                                                       rglpszRQName: new[] { rqName },
+                                                       lpszNewName: newName,
+                                                       promptContinueOnFail: 1);
         }
 
         public async Task OnAfterGlobalSymbolRenamedAsync(string projectPath, IEnumerable<string> filePaths, string rqName, string newName)
         {
-            Requires.NotNull(projectPath, nameof(projectPath));
-            Requires.NotNull(filePaths, nameof(filePaths));
-            Requires.NotNull(rqName, nameof(rqName));
-            Requires.NotNull(newName, nameof(newName));
+            IVsHierarchy? projectHierarchy = await GetProjectHierarchy(projectPath);
+            if (projectHierarchy == null)
+            {
+                return;
+            }
 
-            Instance instance = await WaitForLoadedAsync();
-            await instance.OnAfterGlobalSymbolRenamedAsync(projectPath, filePaths, rqName, newName);
+            if (!(projectHierarchy is IVsHierarchyRefactorNotify refactorNotify))
+            {
+                return;
+            }
+
+            uint[] ids = GetIdsForFiles(projectHierarchy, filePaths).ToArray();
+
+            refactorNotify.OnGlobalSymbolRenamed(cItemsAffected: (uint)ids.Length,
+                                                 rgItemsAffected: ids,
+                                                 cRQNames: (uint)1,
+                                                 rglpszRQName: new[] { rqName },
+                                                 lpszNewName: newName);
+        }
+
+        private async Task<IVsHierarchy?> GetProjectHierarchy(string projectPath)
+        {
+            Project? project = await TryGetProjectFromPathAsync(projectPath);
+            if (project == null)
+            {
+                return null;
+            }
+
+            return await TryGetIVsHierarchyAsync(project);
+        }
+
+        private async Task<Project?> TryGetProjectFromPathAsync(string projectPath)
+        {
+            DTE? dte = await _dte.GetValueAsync();
+            Assumes.NotNull(dte);
+
+            foreach (Project project in dte.Solution.Projects)
+            {
+                string? fullName = null;
+                try
+                {
+                    fullName = project.FullName;
+                }
+                catch (Exception)
+                {
+                    // DTE COM calls can fail for any number of valid reasions.
+                    continue;
+                }
+
+                if (StringComparers.Paths.Equals(fullName, projectPath))
+                {
+                    return project;
+                }
+            }
+
+            return null;
+        }
+
+        private async Task<IVsHierarchy?> TryGetIVsHierarchyAsync(Project project)
+        {
+            IVsSolution? solutionService = await _solutionService.GetValueAsync();
+
+            if (solutionService!.GetProjectOfUniqueName(project.UniqueName, out IVsHierarchy projectHierarchy) == HResult.OK)
+            {
+                return projectHierarchy;
+            }
+
+            return null;
+        }
+
+        private static IEnumerable<uint> GetIdsForFiles(IVsHierarchy projectHierarchy, IEnumerable<string> filePaths)
+        {
+            foreach (string filePath in filePaths)
+            {
+                if (projectHierarchy.ParseCanonicalName(filePath, out uint id) == HResult.OK)
+                {
+                    yield return id;
+                }
+            }
         }
     }
 }
